requirements.txt
Flask
requests
APScheduler
python-levenshtein
python-dateutil
Flask-Login
Flask-WTF